[improve][build] Extract a reusable pulsar.fastutil-minimized-conventions plugin
### Motivation

The fastutil-minimization setup (shadow minimize() seeded from `api` roots, bundle-only
filter, stripped outgoing variants, class-count verification) was inline in
pulsar-client-fastutil-minimized. A second minimization module is coming
(pulsar-broker-fastutil-minimized), so the shared machinery moves into a convention plugin.

### Modifications

- Add build-logic `pulsar.fastutil-minimized-conventions` precompiled script plugin and a
  `FastutilMinimizedExtension`. A consuming module declares its reachability roots as
  `api(project(...))` dependencies and sets `fastutilMinimized { maxRetainedClasses.set(N) }`;
  the plugin handles minimize()/include/extendsFrom-stripping and the verifyMinimizedJar check
  (configuration-cache compatible — the task action captures only Providers).
- Reduce pulsar-client-fastutil-minimized/build.gradle.kts to apply the plugin, declare the
  pulsar-client-original root, and set the 600-class limit.
