+++ pulsar-client-fastutil-minimized/build.gradle.kts
@@ -17,90 +17,22 @@
  * under the License.
  */
 
-// Imported explicitly: the `java` plugin contributes a `java { }` extension accessor,
-// so an unqualified `java.util.zip.ZipFile` would resolve `java` to that extension.
-import java.util.zip.ZipFile
-
-// Produces a jar containing ONLY the classes of the "minimized" dependencies that
-// are reachable from pulsar-client-original (and its full transitive closure) —
-// exactly the classes that end up shaded into the final pulsar-client jars. It is
-// bundled (and relocated) by :pulsar-client-shaded, :pulsar-client-all and
-// :pulsar-client-admin-shaded so the full dependency jars are not shipped.
-//
-// This is the Gradle equivalent of the branch-4.2 Maven module, which uses
-// maven-shade-plugin's <minimizeJar> over {pulsar-client-original, fastutil}.
-//
-// How it works:
-//   * pulsar-client-original is declared with the `api` scope. The Shadow plugin's
-//     minimize() seeds its reachability analysis (UnusedTracker) from the project's
-//     own source classes plus its `api`-scoped jars — so the entire pulsar-client
-//     closure becomes the set of reachability roots. (With `implementation`, or no
-//     source, minimize() has no roots and keeps the full ~12,900-class fastutil jar.)
-//   * Only the minimized libraries are bundled into the jar (the `include` filter
-//     below); pulsar-client-original itself is read purely as a reachability root.
-//   * minimize() then drops every bundled class not reachable from those roots.
+// Jar of just the fastutil classes reachable from the Pulsar client. It is bundled (and
+// relocated) by :pulsar-client-shaded, :pulsar-client-all and :pulsar-client-admin-shaded
+// so the full ~25MB fastutil jar is not shipped. See pulsar.fastutil-minimized-conventions.
 
 plugins {
-    id("pulsar.java-conventions")
-    id("pulsar.shadow-conventions")
+    id("pulsar.fastutil-minimized-conventions")
 }
 
-// Dependencies to minimize, as "group:name" entries. Only the classes from these
-// artifacts that are actually reachable from the pulsar-client closure are kept. Add
-// more entries here when another heavy dependency needs to be minimized the same way.
-val minimizedDependencies: List<String> = listOf(
-    "it.unimi.dsi:fastutil",
-)
-
 dependencies {
-    // `api` (not `implementation`) so minimize() uses pulsar-client-original and its
-    // transitive closure as the reachability roots. Its own classes are not bundled.
+    // Reachability root: pulsar-client-original (its NegativeAcksTracker is the only direct
+    // fastutil consumer on the client side). minimize() seeds from its transitive closure.
     api(project(":pulsar-client-original"))
 }
 
-// The api dependency above is a BUILD-ONLY reachability seed for minimize(). This module
-// ships a self-contained jar of minimized fastutil classes and must not drag
-// pulsar-client-original (or anything else) onto consumers' classpaths, so strip all
-// inherited dependencies from the consumable (outgoing) variants.
-listOf("apiElements", "runtimeElements").forEach { variant ->
-    configurations.named(variant) {
-        setExtendsFrom(emptySet())
-    }
-}
-
-tasks.shadowJar {
-    // Bundle ONLY the minimized libraries; pulsar-client-original is read by minimize()
-    // as a reachability root but is excluded from the output jar.
-    dependencies {
-        minimizedDependencies.forEach { coords -> include(dependency("$coords:.*")) }
-    }
-    // Drop every bundled class not reachable from the api reachability roots above.
-    minimize()
-}
-
-// Verification: the reachable set is ~591 classes; this upper bound is comfortably
-// above it but far below the full fastutil jar (~12,965 classes), so a minimize()
-// regression (e.g. the no-op that ships the whole jar) fails the build. Bump it if
-// new fastutil usage legitimately grows the minimized set past the limit.
-val maxRetainedClasses = 600
-
-val verifyMinimizedJar by tasks.registering {
-    val jarFile = tasks.shadowJar.flatMap { it.archiveFile }
-    val maxClasses = maxRetainedClasses
-    inputs.file(jarFile)
-    doLast {
-        val classCount = ZipFile(jarFile.get().asFile).use { zf ->
-            zf.entries().asSequence().count { it.name.endsWith(".class") }
-        }
-        if (classCount > maxClasses) {
-            throw GradleException(
-                "Minimized jar retained $classCount classes (> $maxClasses) — minimize() is not pruning."
-            )
-        }
-        logger.lifecycle("Minimized fastutil jar OK: $classCount classes retained (limit $maxClasses).")
-    }
-}
-
-tasks.named("check") {
-    dependsOn(verifyMinimizedJar)
+fastutilMinimized {
+    // The reachable set is ~591 classes; fail the build if it grows past this (e.g. if
+    // minimize() regresses and ships the full ~12,965-class jar).
+    maxRetainedClasses.set(600)
 }
